♻ `str_to_bool` 函数放在 `helpers.py` 会产生循环 import
